% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{aggregateDFList}
\alias{aggregateDFList}
\title{aggregateDFList}
\usage{
aggregateDFList(list)
}
\arguments{
\item{list}{a list of same sized dataframes}
}
\description{
Helper function to aggregate and give means of a list of same sized data frames
}
