% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{binaryMatrix}
\alias{binaryMatrix}
\title{binaryMatrix}
\usage{
binaryMatrix(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
result A data frame with 0s for missing data and 1s for present data
}
\description{
A function returning a binary matrix with 0s for missing and 1s for present data
}
