% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{check.miss}
\alias{check.miss}
\title{Title check.miss
check the percent missingness. From juuussi/impute-metabo}
\usage{
check.miss(data, percentage = 0.8)
}
\arguments{
\item{data}{data matrix}
}
\value{
listVar
 list of 3 vectors
 MissingVar     = vector containing the columns numbers of the data matrix with missing values that are less than 80 % NAs
 ExcludedVar    = vector containing the columns numbers of the data matrix with missing values that are more than 80% NAs
 CompleteVar    = vector containing the columns numbers of the data matrix without NAs
}
\description{
Title check.miss
check the percent missingness. From juuussi/impute-metabo
}
