% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{correlationMatrix}
\alias{correlationMatrix}
\title{correlationMatrix
function returning a correlation matrix and NA_correlation matrix}
\usage{
correlationMatrix(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
a list containing a correlation matrix, NA_correlation matrix
}
\description{
correlationMatrix
function returning a correlation matrix and NA_correlation matrix
}
