% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{CullenFrey}
\alias{CullenFrey}
\title{CullenFrey}
\usage{
CullenFrey(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
Function to produce a cullen and frey plot with every feature plotted
}
