% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{detect.MCAR.MNAR.MAR}
\alias{detect.MCAR.MNAR.MAR}
\title{detect.MCAR.MNAR.MAR
from juuussi/impute-metabo}
\usage{
detect.MCAR.MNAR.MAR(
  data,
  MissingVar,
  MAR_MNAR,
  alpha = 0.05,
  percentage = 0.6
)
}
\arguments{
\item{data}{,
data matrix}

\item{MissingVar}{vector containing the columns numbers of the data matrix with missing values that are less than 80 % (use the detect.miss.MNAR.MAR function)}

\item{MAR_MNAR}{vector containing the columns with missing values that have MAR or MNAR missigness (use the detect.miss.MNAR.MAR function)}

\item{alpha}{,
Significance level ,  default is 0.05}

\item{percentage}{}
}
\value{
results
list of  vectors :
MCAR = vector conatining the column numbers of MCAR variables,
MNAR = vector conatining the column numbers of MNAR variables ,
MAR  = vector conatining the column numbers of MAR variables ,
Excluded_marmnar  = vector conatining the column numbers of excluded MAR or MNAR variables that have more than 60% NAs  ,
ExcludedVar       = vector conatining the column numbers of excluded variables that have more than 80% NAs  ,
CompleteVar       = conatining the column numbers of variables without missing values.
}
\description{
detect.MCAR.MNAR.MAR
from juuussi/impute-metabo
}
