% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{detect.miss.MNAR.MAR}
\alias{detect.miss.MNAR.MAR}
\title{detect.miss.MNAR.MAR
It detects if the missigness depends on other varibales
Correlates an indicator matrix (0 not missing , 1 missing) with  the original data that can help determine
if variables tend to be missing together (MAR) or not (MCAR).
Kabacoff, Robert I. R in Action. manning, 2010.
From juuussi/impute-metabo}
\usage{
detect.miss.MNAR.MAR(data, alpha = 0.05, correlation_method)
}
\arguments{
\item{data}{,
matrix with missing values}

\item{alpha, }{significance level ,  default is 0.05}
}
\value{
results
list of vectors :
MissingVar  = vector containing the columns numbers of the data matrix with missing values that are less than 80 % NAs,
PairsCorVar = data frame containing the pairs of correlated variables in the data matrix,
MAR_MNAR    = vector containing the columns numbers of the data matrix with MAR or MNAR missingness
ExcludedVar = vector containing the columns numbers of the data matrix with missing values that are more than 80% NAs
}
\description{
detect.miss.MNAR.MAR
It detects if the missigness depends on other varibales
Correlates an indicator matrix (0 not missing , 1 missing) with  the original data that can help determine
if variables tend to be missing together (MAR) or not (MCAR).
Kabacoff, Robert I. R in Action. manning, 2010.
From juuussi/impute-metabo
}
\examples{
marietta <-detect.miss.MNAR.MAR (miss_data)
}
