% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{detect_missingness_type}
\alias{detect_missingness_type}
\title{detect_missingness_type
Vector conatining the the columns type of missingness. From juuussi/impute-metabo}
\usage{
detect_missingness_type(missigness)
}
\arguments{
\item{missigness}{list of vectors obtained by the detect.MCAR.MNAR.MAR function}
}
\value{
miss_type =  vector conatining the the columns type of missingness
}
\description{
detect_missingness_type
Vector conatining the the columns type of missingness. From juuussi/impute-metabo
}
