% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R
\name{errorEvals}
\alias{errorEvals}
\title{errorEvals}
\usage{
errorEvals(origData, missData, method, imputationResults, simulate_Data = T)
}
\arguments{
\item{method}{string "NRMSE" "NRMSE-SOR", "PCA-P"}

\item{imputationResults}{is specifically a list of imputed dataframes from all the chosen methods, used in SOR}
}
\description{
this will be a function that utilizes all of the different error analysis tools into one nice function
}
