% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R
\name{graphEval}
\alias{graphEval}
\title{graphEval}
\usage{
graphEval(ggplotlist)
}
\arguments{
\item{ggplotlist}{this is a list by missingness ratio and then percents}
}
\value{
a list of plot collections by error evaluation type.
}
\description{
Function that takes in the error result list from rearrangeList which is from simulateEngine followed by aggregateDF and outputs multiplots for each missingness
ratio and makes graphs for each error type in the multiplot
}
