% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccEvaluation.R
\name{ICC_Change_Counts}
\alias{ICC_Change_Counts}
\title{ICC_Change_Counts}
\usage{
ICC_Change_Counts(iccImputed, iccComparison)
}
\arguments{
\item{iccImputed}{imputed ICCs vector}

\item{iccComparison}{vector of ICCs of the comparison method (eg zero imputation)}
}
\value{

}
\description{
Counts of ICC changes from one class of Koo et al. to another after imputation
}
