% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccEvaluation.R
\name{ICC_Counts}
\alias{ICC_Counts}
\title{ICC Counts}
\usage{
ICC_Counts(iccMatrix)
}
\arguments{
\item{iccMatrix}{matrix of ICCs for all imputation methods}
}
\value{

}
\description{
Function to count ICCs in a certain classification by Koo et al. Excellent is >.9, good is from .75 to .9, moderate from .5 to .75, poor from 0 to 0.5 and inconclusive
below 0
}
