% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R, R/iccEvaluation.R
\name{ICCformatting}
\alias{ICCformatting}
\title{ICCformatting}
\usage{
ICCformatting(data, reps)

ICCformatting(data, reps)
}
\arguments{
\item{reps}{replicate number}
}
\value{
matrix

matrix
}
\description{
Function that formats a dataframe or matrix into a dataframe where there are at least 2 factor levels that have non
missing data.

Function that formats a dataframe or matrix into a dataframe where there are at least (replicate # - 1) factor levels that have non
missing data.
}
