% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{Impute}
\alias{Impute}
\title{Title Impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods}
\usage{
Impute(data, method, local = TRUE, reps)
}
\arguments{
\item{data}{data matrix with simulated data}

\item{method}{the imputation method Note within replicate methods are preceded by 'R', eg 'RMIN'
method=c('RF', 'BPCA', 'QRILC', 'GSIMP', 'RHM','RMEAN', 'RMEDIAN', 'RMIN','RZERO', 'RRF',
'RGSIMP', 'RQRILC','RBPCA','min','halfmin', 'mean', 'median', 'zero')}

\item{local}{a boolean to determine if local rep_impute method is to be used, default to true.}

\item{reps}{the number of replicate groups
simulated data that doesn't require log preprocessing.}
}
\value{
results_data the matrix containing the imputed data
}
\description{
Title Impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods
}
\examples{
imputed_data <- impute(data=miss_data, methods=imputation_methods, local=T, reps=3)
#'####################################################
}
