% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{missingCounter}
\alias{missingCounter}
\title{missingCounter}
\usage{
missingCounter(vector)
}
\arguments{
\item{vector}{A vector of values}
}
\description{
A function that counts missingness in a vector
}
