% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{missingProportions}
\alias{missingProportions}
\title{missingProportions}
\usage{
missingProportions(varStats)
}
\arguments{
\item{varStats}{a datafram of variable statistics as given by variableStatistics function}
}
\value{
result a data frome of the results
}
\description{
A function that estimates missingness proportions across a data set
}
