% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{multi_impute}
\alias{multi_impute}
\title{copied from WandeRum/GSimp}
\usage{
multi_impute(
  data_miss,
  iters_each = 100,
  iters_all = 20,
  initial = "qrilc",
  lo = -Inf,
  hi = "min",
  n_cores = 1,
  imp_model = "glmnet_pred",
  gibbs = data.frame(row = integer(), col = integer())
)
}
\description{
copied from WandeRum/GSimp
}
