% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{plotResults}
\alias{plotResults}
\title{plotresults}
\usage{
plotResults(results, missRatios, missMax, missMin, missInc, simIter)
}
\arguments{
\item{results}{from simulate engine}

\item{missRatios}{from sim engine}

\item{missMax}{from simulate engine}

\item{missMin}{from simulate engine}

\item{missInc}{from simulae engine}

\item{simIter}{from simulate engine}
}
\value{
a ggplot formatted list
}
\description{
The purpose of this functino is to take the results from simulateEngine and output a list of data frames
that are useable by ggPlots
}
