% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{scale_recover}
\alias{scale_recover}
\title{Scale and recover -------------------------------------------------------
function copied from GSimp by WandeRum/GSimp}
\usage{
scale_recover(data, method = "scale", param_df = NULL)
}
\description{
Scale and recover -------------------------------------------------------
function copied from GSimp by WandeRum/GSimp
}
