% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R
\name{simEval}
\alias{simEval}
\title{simEval}
\usage{
simEval(
  origData,
  missData,
  impData,
  methods,
  outcome = NULL,
  imputationResults
)
}
\arguments{
\item{origData}{the dataset prior to simulating missingness}

\item{missData}{the dataset with missingness applied}

\item{impData}{the imputed dataset}

\item{methods}{a list of evaluation methods to be utilized}

\item{outcome}{is a grouping list for supervized learning to use PLS or pearson}

\item{imputationResults}{}
}
\value{
results a dataframe displaying NRMSE, PCA-Procustes, PLS Procrustes and Student's T test and Pearson Correl
}
\description{
simEval
}
