% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate_missingness}
\alias{simulate_missingness}
\title{Title simulate_missingness
Uses the simulated data to create the different types of missingness (MCAR,MNAR,MAR)
using different percentages of missigness}
\usage{
simulate_missingness(
  data,
  mcar = 0,
  mar = 0,
  mnar = 0,
  mnar.type = "left",
  mar.type = "left"
)
}
\arguments{
\item{data}{matrix with simulated data}

\item{mcar}{percenatge of missigness in Missing Completly at Random}

\item{mar}{percenatge of missigness in Missing at Random}

\item{mnar}{percenatge of missignessMisiing not at Random}

\item{mnar.type}{type of truncation 'left' or right}

\item{mar.type}{type of truncation}
}
\value{
simulated_data
}
\description{
Title simulate_missingness
Uses the simulated data to create the different types of missingness (MCAR,MNAR,MAR)
using different percentages of missigness
}
