% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulateEngine}
\alias{simulateEngine}
\title{simulateEngine}
\usage{
simulateEngine(
  data,
  simIter,
  simMissIter,
  missMax,
  missMin,
  missInc,
  missRatios,
  methodsImp,
  methodsEval,
  simulate_Data = T,
  reps
)
}
\arguments{
\item{data}{the original data set}

\item{simIter}{this is the number of unique simulation matrices to be made from the covMat of the origData}

\item{simMissIter}{this is the number of simulated missing data matrices to be created for a given missingness
proportion}

\item{missMax}{the highest amount of missingness wanted in the sim}

\item{missMin}{the min amount of missingness wanted in the sim}

\item{missInc}{the amount of missingness to increment by}

\item{missRatios}{this is a list of triplets defining the ratio of MCAR, MAR, MNAR in form c(0,0.5, 0.5) would
specify that we want a split of 50-50 MAR MNAR for each missingness level}

\item{methodsImp}{the imputation methods to be utilized.}

\item{medthodsEval}{the methods of error evaluation to be utilized.}
}
\value{
a collection of dataframes containing the errors of imp methods for each eval method listed by
proportion of missingness and total percent of missingness.
}
\description{
This function is going to be the main engine that actually runs the simulations and tracks the data and the
error calculations at differing levels of missingness, ratios of missingness and iterations of the sim.
}
