% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{single_impute_iters}
\alias{single_impute_iters}
\title{copied from WandeRum/GSimp}
\usage{
single_impute_iters(
  x,
  y,
  y_miss,
  y_real = NULL,
  imp_model = "glmnet_pred",
  lo = -Inf,
  hi = Inf,
  iters_each = 100,
  gibbs = c()
)
}
\description{
copied from WandeRum/GSimp
}
