% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R
\name{errorEvals}
\alias{errorEvals}
\title{errorEvals}
\usage{
errorEvals(origData, missData, method, outcome = NULL, imputationResults)
}
\arguments{
\item{method}{string "NRMSE" "NRMSE SOR", "PCA Procrustes", "PLS Procrustes", "Pearson Corr"}

\item{outcome}{if there are groups and we want to use PLS-DA which is a supervised algo}

\item{imputationResults}{is specifically a list of imputed dataframes from all the chosen methods, used in SOR}
}
\description{
this will be a function that utilizes all of the different error analysis tools into one nice function
}
