% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccEvaluation.R
\name{ICC_Change_Plot}
\alias{ICC_Change_Plot}
\title{ICC Change Plot}
\usage{
ICC_Change_Plot(iccMeasure, methods)
}
\arguments{
\item{iccMeasure}{matrix of a measure of ICC change from ICC Eval}

\item{methods}{list of methods of imputation with baseline comparison matrix last}
}
\value{
a ggplot of measure of icc change with CIs
}
\description{
ICC Change Plot
}
