% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccEvaluation.R
\name{ICC_Scatter_Plot}
\alias{ICC_Scatter_Plot}
\title{ICC Scatter Plot}
\usage{
ICC_Scatter_Plot(rawData, reps, iccImputed, iccComparison, plotTitle)
}
\arguments{
\item{rawData}{original dataset}

\item{reps}{number of replicate groups}

\item{iccImputed}{ICCs of the data after an imputation method}

\item{iccComparison}{ICCs of the comparison data (eg zero imputed matrix)}

\item{plotTitle}{string}
}
\value{
result list of the scatter plot data and the ggplot
}
\description{
ICC Scatter Plot
}
