% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorEvaluation.R, R/iccEvaluation.R
\name{iccEval}
\alias{iccEval}
\title{iccEval}
\usage{
iccEval(origData, reps, imputed, methods)

iccEval(origData, reps, imputed, methods)
}
\arguments{
\item{reps}{number of replicates}

\item{imputed}{this is a list of imputed dataframes}

\item{methods}{this is a vector of all the imputation methods used}

\item{groups}{this is a factor containing group levels}
}
\value{
a list of various measures

a list of various measures
}
\description{
a function to compare ICC of original data to imputed data. MAKE SURE THAT THE LAST IN METHODS VECTOR is zero imputation

a function to compare ICC of original data to imputed data. MAKE SURE THAT THE LAST IN METHODS VECTOR is the comparison imputation method (zero usually)
}
