% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute}
\alias{impute}
\title{Title impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods}
\usage{
impute(data, methods, local = TRUE, reps)
}
\arguments{
\item{data}{data matrix with simulated data}

\item{methods}{vector containing the method names. Note GSimp_Real is for real non-neg data. GSimp_Sim is for already}

\item{local}{a boolean to determine if local rep_impute method is to be used, default to true.}

\item{reps}{the number of replicate groups
simulated data that doesn't require log preprocessing.}
}
\value{
the matrix containing the imputed data
}
\description{
Title impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods
}
\examples{
imputed_data <- impute(data=miss_data, methods=imputation_methods)
#'####################################################
}
