% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R
\name{impute}
\alias{impute}
\title{Title impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods}
\usage{
impute(data, method, local = TRUE, reps)
}
\arguments{
\item{data}{data matrix with simulated data}

\item{method}{the imputation method Note GSimp_Real is for real non-neg data. GSimp_Sim is for already.
method=c('RF', "BPCA", 'QRILC', 'GSimp_Log', 'Rep_Imp_HM',"Rep_Imp_mean", "Rep_Imp_median", "Rep_Imp_min","Rep_Zero", "Rep_Imp_RF",
"Rep_Imp_GSimp", "Rep_Imp_QRILC","Rep_Imp_BPCA",'min',"halfmin", 'mean', 'median', 'zero')}

\item{local}{a boolean to determine if local rep_impute method is to be used, default to true.}

\item{reps}{the number of replicate groups
simulated data that doesn't require log preprocessing.}
}
\value{
results_data the matrix containing the imputed data
}
\description{
Title impute
This functions contains different imputation methods and imputes the data with all
the different imputation methods
}
\examples{
imputed_data <- impute(data=miss_data, methods=imputation_methods, local=T, reps=3)
#'####################################################
}
