% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{simulate}
\alias{simulate}
\title{simulate
function to simulate a similar data set with similar correlation matrix.}
\usage{
simulate(rownum, colnum, corMat)
}
\arguments{
\item{rownum}{number of rows of the original data matrix}

\item{colnum}{number of columns of the original data matrix}

\item{corMat}{correlation matrix.}

\item{meanval}{mean for normalization}

\item{sdval}{standard deviation for normalization}
}
\value{
a matrix of the original size of your data matrix with simulated values (normalized)
}
\description{
simulate
function to simulate a similar data set with similar correlation matrix.
}
