% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{variableStatistics}
\alias{variableStatistics}
\title{variableStatistics}
\usage{
variableStatistics(
  data,
  correlation_method = "pearson",
  gof_method = "kolmogorov"
)
}
\arguments{
\item{data}{A data frame (rows as samples, columns as variables)}

\item{correlation_method}{the method to compute correlations 'pearson' (default), 'spearman', 'kendall'}

\item{gof_method}{the method to compute goodness of fit with left-censored data 'kolmogorov' (default), 'cucconi'}
}
\value{
result A list containing [[1]] a matrix of: percent missingness, quartile cutoffs, mean, missingness type, variance, distribution for each variable;
[[2]] a matrix containing missingness percents by missingness type and [[3]] the overall missingness percentage
}
\description{
A function that does various statistics on missingness per variables
}
