% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_Eval_functions.R
\name{variableStatistics}
\alias{variableStatistics}
\title{variableStatistics}
\usage{
variableStatistics(data)
}
\arguments{
\item{data}{A data frame (rows as samples, columns as variables)}
}
\value{
result A matrix containing: percent missingness, quartile cutoffs, mean, missingness type, variance, distribution
}
\description{
A function that does various statistics on missingness per variables
}
